0725demo/requirements.txt
google-genai
python-dotenv
Pillow
tqdm
Unidecode